requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
requests
anthropic
gunicorn
whitenoise
psycopg2-binary
